[refactor] command 패키지 하위에 testscript, testshell 패키지 위치하도록 패키지 구조 변경
