--- shell/src/main/java/com/samsung/Command.java
@@ -1,5 +1,0 @@
-package com.samsung;
-
-public interface Command {
-    void execute(String[] cmdArgs);
-}
