+++ shell/src/main/java/com/samsung/command/Command.java
@@ -1,0 +1,5 @@
+package com.samsung.command;
+
+public interface Command {
+    void execute(String[] cmdArgs);
+}
