+++ shell/src/main/java/com/samsung/command/testshell/ReadCommand.java
@@ -1,0 +1,35 @@
+package com.samsung.command.testshell;
+
+import com.samsung.command.Command;
+
+public class ReadCommand implements Command {
+
+    private final TestShellManager testShellManager;
+
+    public ReadCommand(TestShellManager testShellManager) {
+        this.testShellManager = testShellManager;
+    }
+
+    @Override
+    public void execute(String[] cmdArgs) {
+        // 명령어 길이 체크
+        if (cmdArgs.length != 2) {
+            throw new RuntimeException("INVALID COMMAND");
+        }
+
+        try {
+            Integer index = Integer.parseInt(cmdArgs[1]);
+
+            // null 체크?
+
+            // lba 범위 검증
+            if (index < 0 || index > 99) {
+                throw new RuntimeException("INVALID COMMAND");
+            }
+
+            testShellManager.read(index);
+        } catch (NumberFormatException e) {
+            throw new RuntimeException("INVALID COMMAND"); // 숫자 아니면 Exception
+        }
+    }
+}
