--- shell/src/test/java/testscript/command/TestScript3CommandTest.java
@@ -1,29 +1,0 @@
-package testscript.command;
-
-import com.samsung.testscript.ScriptManager;
-import com.samsung.testscript.command.TestScript2Command;
-import com.samsung.testscript.command.TestScript3Command;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-import static org.mockito.Mockito.verify;
-
-@ExtendWith(MockitoExtension.class)
-public class TestScript3CommandTest {
-    @Mock
-    ScriptManager scriptManager;
-    @InjectMocks
-    TestScript3Command testScript3Command;
-
-    @Test
-    @DisplayName("TestScript3 실행 테스트")
-    void callTestScript3Command(){
-        testScript3Command.execute(new String[1]);
-
-        verify(scriptManager).testScript3();
-    }
-}
